Integrate Streams membership manager with stream thread (#17795)
A stream thread needs to integrate with the Streams membership manager to receive an assignment. This happens through callbacks that are set in the streams assignment interface. The Streams membership manager calls three callbacks:
- onTasksRevoked that revokes tasks assigned to the stream thread
- onTaskAssigned that assigns tasks to the stream thread
- onAllTasksLost that releases all tasks of a stream thread and closes them dirty.

Initially, we thought we could avoid onTasksRevoked by just considering revoked tasks in onTaskAssigned. However, it turns out we cannot avoid onTasksRevoked because if we do that we would need to set the new assignment in the subscription state before or after the call to onTaskAssigned. If we set the new assignment in the subscription state before we call onTaskAssigned, commits of the revoked tasks will fail because the subscription state would not contain the partitions of the revoked tasks anymore. If we set the new assignment in the subscription state after we call onTaskAssigned, we cannot safely pause the partitions of the assigned task for restoration because we do not know when the new assignment is actually set in the subscription state since that happens on the background thread.